+++ docs/history.md
@@ -7,6 +7,8 @@ V2.2.4
 - *Unreleased*
 
 * add list of op commands that will not produce changes
+* Add list of type commands that will not produce changes
+* Return output from commands in `panos_type_cmd`
 
 V2.2.3
 ------